Fix MSVC warnings
Input.c(103): warning C4267: 'function': conversion from 'size_t' to 'UINT', possible loss of data

Layout.c(96): warning C4267: 'function': conversion from 'size_t' to 'DWORD', possible loss of data
Layout.c(102): warning C4267: 'function': conversion from 'size_t' to 'DWORD', possible loss of data
Layout.c(108): warning C4267: 'function': conversion from 'size_t' to 'DWORD', possible loss of data
Layout.c(115): warning C4267: 'function': conversion from 'size_t' to 'DWORD', possible loss of data

Str.c(23): warning C4267: 'function': conversion from 'size_t' to 'UINT', possible loss of data
Str.c(51): warning C4477: 'sscanf_s' : format string '%c' requires an argument of type 'unsigned int', but variadic argument 9 has type 'size_t'
Str.c(65): warning C4477: 'sscanf_s' : format string '%c' requires an argument of type 'unsigned int', but variadic argument 4 has type 'size_t'
Str.c(78): warning C4477: 'sscanf_s' : format string '%c' requires an argument of type 'unsigned int', but variadic argument 5 has type 'size_t'

+++ Input.c
@@ -100,7 +100,7 @@ bool disableLayout(const uint16_t device_id, const uint16_t lang_id)
 	UINT num = EnumEnabledLayoutOrTip(NULL, NULL, NULL, NULL, 0);
 	const size_t size = sizeof(LayoutOrTipProfile) * num;
 	LayoutOrTipProfile *profiles = malloc(size);
-	num = EnumEnabledLayoutOrTip(NULL, NULL, NULL, profiles, size);
+	num = EnumEnabledLayoutOrTip(NULL, NULL, NULL, profiles, (UINT)size);
 
 	for (UINT i = 0; i < num; ++i) {
 		const LayoutOrTipProfile *profile = &profiles[i];
